# BUILD : 1.7.15 (1136) # BUGS  : 578 # NOTES : fixed bsd compiler warning
git-svn-id: svn://svn.anope.org/anope/trunk@1136 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@859 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/protocol/inspircd.c
@@ -604,7 +604,7 @@ int anope_event_idle(char *source, int ac, char **av)
 {
 	if (ac == 1)
 	{
-		send_cmd(av[0],"IDLE %s %ld 0",source,time(NULL));
+		send_cmd(av[0],"IDLE %s %ld 0",source,(long int) time(NULL));
 	}
 	return MOD_CONT;
 }
